Common: Added generic ContinuousSimulator

+++ hio-dotnet.Common/Helpers/TimeHelpers.cs
@@ -35,7 +35,7 @@ namespace hio_dotnet.Common.Helpers
         /// </summary>
         /// <param name="dateTime"></param>
         /// <returns></returns>
-        public static double DateTimeToUnixTimestamp(DateTime dateTime)
+        public static long DateTimeToUnixTimestamp(DateTime dateTime)
         {
             DateTime unixStart = new DateTime(1970, 1, 1, 0, 0, 0, 0, System.DateTimeKind.Utc);
             long unixTimeStampInTicks = (dateTime.ToUniversalTime() - unixStart).Ticks;
